calculate aggregates on each new measurement
Rather than keeping a complete set of values recorded for individual
metrics, keep running statistics (count, sum, sum_squares, etc) and push
those to the Librato server. This should have negligible impact when
writing single-valued metrics (and those with a small number of values)
but should have a huge impact when writing metrics with a large number
of values.