Stop testing for Node 0.x releases
These are extremely outdated, so there's probably no good reason to keep
supporting them. We could attempt to install even lower version of nock,
but that shows vulnerability alerts in development, which while not
very important, seems scary and confusing.

We could also drop the support for Node 4 and install the latest version
of nock.