Address audit warnings by bumping mocha version
Mocha 3 started failing tests that were relying on the test callback
function (`done`), but were returning Promises at the same time.
`requestretry()` returns a Promise, and so does `Client.send`.

So we had two options:

1. stop calling the `done()` callbacks in tests
2. stop returning Promises from `it()` in tests

Option 1 was not viable, because the Promise returned by `Client.send()`
is not usable. It's the same promise that's returned by
`requestretry()`, which will throw "A callback was provided but waiting
a promise, use only one pattern" if `requestretry()` was given a
callback and the Promise's `then()` was also called.

For option 2 we could either change the return value of `Client.send()`,
which is probably be safe, but breaks backwards compatibility, or adjust
the return value from `it()` directly.