fix: more gracefully handle time travel
If the clock date moves backwards, say from 2020 to 2010, the librato-node worker
schedules a `setTimeout()` callback for ten years in the future, which causes node to log
a warning due to 32-bit signed integer overflow.

In practice, the behavior isn't great - that callback will basically never fire. We could
fix this by updating the logic to handle this somewhat gracefully, but if your clock is
moving backwards, you have other issues to deal with anyway. So this just focuses on
suppressing the warning by ensuring we don't pass a massive timeout to `setTimeout()`.

I arbitrarily capped it the period (default 60s). This means the behavior in practice is
still not great - it will schedule another check every 60s for eternity. But that seems
about the same as before, to me.